backend/requirements.txt
fastapi
uvicorn
pydantic
groq
requests
PyPDF2
python-multipart
python-dotenv
